Moving test cases to they correspondig package
