+++ src/test/java/giis/eshopcontainers/e2e/functional/tests/api/BasketAPITests.java
@@ -1,4 +1,4 @@
-package giis.eshopcontainers.e2e.functional.tests;
+package giis.eshopcontainers.e2e.functional.tests.api;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonArray;
